app_server/requirements.txt
fastapi
uvicorn[standard]
aerospike
kafka-python